Split contributor wall case from SVG templates.
